+++ meinberlin/apps/dashboard2/__init__.py
@@ -1,8 +1,8 @@
 from .contents import content
 from .components import DashboardComponent
-from .components import ModuleFormComponent
-from .components import ModuleFormSetComponent
-from .components import ProjectFormComponent
+from .components.forms import ModuleFormComponent
+from .components.forms import ModuleFormSetComponent
+from .components.forms import ProjectFormComponent
 
 __all__ = ['content', 'DashboardComponent',
            'ModuleFormComponent', 'ModuleFormSetComponent',
